refactoring request and response params
