+++ relay_rpc/src/rpc/params/session.rs
@@ -1,21 +1,6 @@
 use {
-    super::{
-        session_delete::SessionDeleteRequest,
-        session_event::SessionEventRequest,
-        session_extend::SessionExtendRequest,
-        session_propose::{SessionProposeRequest, SessionProposeResponse},
-        session_request::SessionRequestRequest,
-        session_settle::SessionSettleRequest,
-        session_update::SessionUpdateRequest,
-    },
-    crate::rpc::{
-        params::{IrnMetadata, ParamsError, RelayProtocolHelpers, RelayProtocolMetadata},
-        ErrorData,
-    },
-    paste::paste,
     regex::Regex,
     serde::{Deserialize, Serialize},
-    serde_json::Value,
     std::{
         collections::{BTreeMap, BTreeSet},
         ops::Deref,
@@ -241,166 +226,6 @@ pub struct SettleNamespace {
     pub accounts: BTreeSet<String>,
     pub methods: BTreeSet<String>,
     pub events: BTreeSet<String>,
-}
-
-// Convenience macro to de-duplicate implementation for different parameter
-// sets.
-macro_rules! impl_relay_protocol_metadata {
-    ($param_type:ty,$meta:ident) => {
-        paste! {
-            impl RelayProtocolMetadata for $param_type {
-                fn irn_metadata(&self) -> IrnMetadata {
-                    match self {
-                        [<$param_type>]::SessionPropose(_) => $crate::rpc::params::session_propose::[<IRN_ $meta:upper _METADATA>],
-                        [<$param_type>]::SessionSettle(_) => $crate::rpc::params::session_settle::[<IRN_ $meta:upper _METADATA>],
-                        [<$param_type>]::SessionRequest(_) => $crate::rpc::params::session_request::[<IRN_ $meta:upper _METADATA>],
-                        [<$param_type>]::SessionUpdate(_) => $crate::rpc::params::session_update::[<IRN_ $meta:upper _METADATA>],
-                        [<$param_type>]::SessionDelete(_) => $crate::rpc::params::session_delete::[<IRN_ $meta:upper _METADATA>],
-                        [<$param_type>]::SessionEvent(_) => $crate::rpc::params::session_event::[<IRN_ $meta:upper _METADATA>],
-                        [<$param_type>]::SessionExtend(_) => $crate::rpc::params::session_extend::[<IRN_ $meta:upper _METADATA>],
-                        [<$param_type>]::SessionPing(_) => $crate::rpc::params::session_ping::[<IRN_ $meta:upper _METADATA>],
-                    }
-                }
-            }
-        }
-    }
-}
-
-// Convenience macro to de-duplicate implementation for different parameter
-// sets.
-macro_rules! impl_relay_protocol_helpers {
-    ($param_type:ty) => {
-        paste! {
-            impl RelayProtocolHelpers for $param_type {
-                type Params = Self;
-
-                fn irn_try_from_tag(value: Value, tag: u32) -> Result<Self::Params, ParamsError> {
-                    match tag {
-                        tag if tag == $crate::rpc::params::session_propose::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionPropose(serde_json::from_value(value)?))
-                        }
-                        tag if tag == $crate::rpc::params::session_settle::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionSettle(serde_json::from_value(value)?))
-                        }
-                        tag if tag == $crate::rpc::params::session_request::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionRequest(serde_json::from_value(value)?))
-                        }
-                        tag if tag == $crate::rpc::params::session_delete::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionDelete(serde_json::from_value(value)?))
-                        }
-                        tag if tag == $crate::rpc::params::session_extend::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionExtend(serde_json::from_value(value)?))
-                        }
-                        tag if tag == $crate::rpc::params::session_update::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionUpdate(serde_json::from_value(value)?))
-                        }
-                        tag if tag == $crate::rpc::params::session_event::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionEvent(serde_json::from_value(value)?))
-                        }
-                        tag if tag == $crate::rpc::params::session_event::IRN_RESPONSE_METADATA.tag => {
-                            Ok(Self::SessionPing(serde_json::from_value(value)?))
-                        }
-                        _ => Err(ParamsError::ResponseTag(tag)),
-                    }
-                }
-            }
-        }
-    };
-}
-
-/// Sign API request parameters.
-///
-/// https://specs.walletconnect.com/2.0/specs/clients/sign/rpc-methods
-/// https://specs.walletconnect.com/2.0/specs/clients/sign/data-structures
-#[derive(Debug, Serialize, Eq, Deserialize, Clone, PartialEq)]
-#[serde(tag = "method", content = "params")]
-pub enum SessionRequestParams {
-    #[serde(rename = "wc_sessionPropose")]
-    SessionPropose(SessionProposeRequest),
-    #[serde(rename = "wc_sessionSettle")]
-    SessionSettle(SessionSettleRequest),
-    #[serde(rename = "wc_sessionUpdate")]
-    SessionUpdate(SessionUpdateRequest),
-    #[serde(rename = "wc_sessionExtend")]
-    SessionExtend(SessionExtendRequest),
-    #[serde(rename = "wc_sessionRequest")]
-    SessionRequest(SessionRequestRequest),
-    #[serde(rename = "wc_sessionEvent")]
-    SessionEvent(SessionEventRequest),
-    #[serde(rename = "wc_sessionDelete")]
-    SessionDelete(SessionDeleteRequest),
-    #[serde(rename = "wc_sessionPing")]
-    SessionPing(()),
-}
-
-impl_relay_protocol_metadata!(SessionRequestParams, request);
-
-/// https://www.jsonrpc.org/specification#response_object
-///
-/// JSON RPC 2.0 response object can either carry success or error data.
-/// Please note, that relay protocol metadata is used to disambiguate the
-/// response data.
-///
-/// For example:
-/// `RelayProtocolHelpers::irn_try_from_tag` is used to deserialize an opaque
-/// response data into the typed parameters.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub enum ResponseParams {
-    /// A response with a result.
-    #[serde(rename = "result")]
-    Success(Value),
-
-    /// A response for a failed request.
-    #[serde(rename = "error")]
-    Err(Value),
-}
-
-/// Typed success response parameters.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum SessionResponseParamsSuccess {
-    SessionPropose(SessionProposeResponse),
-    SessionSettle(bool),
-    SessionUpdate(bool),
-    SessionExtend(bool),
-    SessionRequest(bool),
-    SessionEvent(bool),
-    SessionDelete(bool),
-    SessionPing(bool),
-}
-impl_relay_protocol_metadata!(SessionResponseParamsSuccess, response);
-impl_relay_protocol_helpers!(SessionResponseParamsSuccess);
-
-impl TryFrom<SessionResponseParamsSuccess> for ResponseParams {
-    type Error = ParamsError;
-
-    fn try_from(value: SessionResponseParamsSuccess) -> Result<Self, Self::Error> {
-        Ok(Self::Success(serde_json::to_value(value)?))
-    }
-}
-
-/// Typed error response parameters.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(untagged)]
-pub enum SessionResponseParamsError {
-    SessionPropose(ErrorData),
-    SessionSettle(ErrorData),
-    SessionUpdate(ErrorData),
-    SessionExtend(ErrorData),
-    SessionRequest(ErrorData),
-    SessionEvent(ErrorData),
-    SessionDelete(ErrorData),
-    SessionPing(ErrorData),
-}
-impl_relay_protocol_metadata!(SessionResponseParamsError, response);
-impl_relay_protocol_helpers!(SessionResponseParamsError);
-
-impl TryFrom<SessionResponseParamsError> for ResponseParams {
-    type Error = ParamsError;
-
-    fn try_from(value: SessionResponseParamsError) -> Result<Self, Self::Error> {
-        Ok(Self::Err(serde_json::to_value(value)?))
-    }
 }
 
 /// Trims json of the whitespaces and newlines.
